refactor: split depth pipeline into camera-agnostic core + ZED/RealSense sources
